+++ lib/view/browsing_view.dart
@@ -2,7 +2,7 @@ import 'package:cool_urban_spaces/controller/map_data_controller.dart';
 import 'package:cool_urban_spaces/controller/settings_controller.dart';
 import 'package:cool_urban_spaces/controller/enum/sorting_type.dart';
 import 'package:cool_urban_spaces/model/suggestion.dart';
-import 'package:cool_urban_spaces/view/widgets/utils_view.dart';
+import 'package:cool_urban_spaces/view/widgets/utils_widget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
