+++ lib/view/info_suggestion_view.dart
@@ -2,8 +2,8 @@ import 'package:cool_urban_spaces/controller/chat_controller.dart';
 import 'package:cool_urban_spaces/controller/map_data_controller.dart';
 import 'package:cool_urban_spaces/controller/profile_controller.dart';
 import 'package:cool_urban_spaces/model/suggestion.dart';
-import 'package:cool_urban_spaces/view/widgets/urban_map_view.dart';
-import 'package:cool_urban_spaces/view/widgets/utils_view.dart';
+import 'package:cool_urban_spaces/view/widgets/urban_map_widget.dart';
+import 'package:cool_urban_spaces/view/widgets/utils_widget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_ui/flutter_chat_ui.dart';
